Refaktorering av obstacle controller og account controller
