+++ Luftfartshinder/Luftfartshinder/Controllers/Account/AccountController.cs
@@ -14,7 +14,7 @@ namespace Luftfartshinder.Controllers.Account
     /// <summary>
     /// Controller for handling user authentication, registration, and account management.
     /// </summary>
-    public class AccountController : Controller
+    public partial class AccountController : Controller
     {
         private readonly UserManager<ApplicationUser> userManager;
         private readonly SignInManager<ApplicationUser> signInManager;
@@ -251,139 +251,7 @@ namespace Luftfartshinder.Controllers.Account
             return RedirectToAction("Login", "Account");
         }
 
-        /// <summary>
-        /// Displays the user dashboard with reports based on user role.
-        /// </summary>
-        [HttpGet]
-        [Authorize]
-        public async Task<IActionResult> Dashboard()
-        {
-            if (User.IsInRole("FlightCrew") && !User.IsInRole("SuperAdmin"))
-            {
-                ViewData["LayoutType"] = "ipad";
-            }
-            else
-            {
-                ViewData["LayoutType"] = "pc";
-            }
-            var reports = await GetReportsForUserAsync();
-            return View(reports);
-        }
 
-        /// <summary>
-        /// Displays obstacles and reports for FlightCrew users.
-        /// </summary>
-        [Authorize]
-        public async Task<IActionResult> FlightCrewObstacles()
-        {
-            ViewData["LayoutType"] = "ipad";
-
-            var user = await userManager.GetUserAsync(User);
-            if (user == null)
-            {
-                return Challenge();
-            }
-
-            if (user.OrganizationId == 0)
-            {
-                return Forbid();
-            }
-
-            var organization = await organizationRepository.GetById(user.OrganizationId);
-            if (organization == null)
-            {
-                return NotFound();
-            }
-
-            var obstacles = await obstacleRepository.GetByOrgId(organization.Id);
-            var reports = await reportRepository.GetByOrgId(organization.Id);
-
-            var viewModel = new OrgDataViewModel
-            {
-                Organization = organization,
-                Obstacles = obstacles,
-                Reports = reports
-            };
-
-            return View("FlightCrewObstacles", viewModel);
-        }
-
-        // Private helper methods
-
-        private UserRegisterViewModel CreateEmptyUserRegisterViewModel()
-        {
-            return new UserRegisterViewModel
-            {
-                FirstName = string.Empty,
-                LastName = string.Empty,
-                Email = string.Empty,
-                Username = string.Empty,
-                Password = string.Empty,
-                SelectedRole = string.Empty,
-                ConfirmPassword = string.Empty
-            };
-        }
-
-        private void AddValidationErrorsToModelState(List<string> errors)
-        {
-            foreach (var error in errors)
-            {
-                if (error.Contains("Email"))
-                {
-                    ModelState.AddModelError("Email", error);
-                }
-                else if (error.Contains("Username"))
-                {
-                    ModelState.AddModelError("Username", error);
-                }
-                else
-                {
-                    ModelState.AddModelError("", error);
-                }
-            }
-        }
-
-        private void AddIdentityErrorsToModelState(IdentityResult? result)
-        {
-            if (result == null) return;
-
-            foreach (var error in result.Errors)
-            {
-                ModelState.AddModelError("", error.Description);
-            }
-        }
-
-        private async Task<IActionResult> RedirectToDashboardByRoleAsync(ApplicationUser? user)
-        {
-            if (user == null)
-            {
-                return RedirectToAction("Tutorial", "Home");
-            }
-
-            var roles = await userManager.GetRolesAsync(user);
-
-            if (roles.Contains("SuperAdmin") || roles.Contains("Registrar"))
-            {
-                return RedirectToAction("Dashboard");
-            }
-
-            if (roles.Contains("FlightCrew"))
-            {
-                return RedirectToAction("Index", "Home");
-            }
-
-            return RedirectToAction("Tutorial", "Home");
-        }
-
-        private async Task<List<Report>> GetReportsForUserAsync()
-        {
-            if (User.IsInRole("Registrar") || User.IsInRole("SuperAdmin"))
-            {
-                return await reportRepository.GetAllAsync();
-            }
-
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-            return await accountRepository.GetUserReports(userId);
-        }
+        
     }
 }
